Create blackjack.py
Blackjack is a game where the player plays against a dealer. The player is dealt 2 cards from a deck of cards, wherein each card has a value, number cards have face value whereas face cards have a value of 10, the ace can be assigned a value of 1 or 11 depending on the situation. The objective of the game is for the player to beat the dealer, which he can do so by getting a total value higher than the dealer's value.